Create and send reset code

+++ src/dependencies.php
@@ -170,7 +170,7 @@ $container['mailer'] = function () {
     if ( getenv('MAIL_TRANSPORT') == "SENDMAIL" ) {
         $transport = new Swift_SendmailTransport('/usr/sbin/sendmail -bs');
     } elseif ( getenv('MAIL_TRANSPORT') == "SMTP" ) {
-        $transport = new Swift_SmtpTransport(getenv('SMTP_SERVER'), getenv('SMTP_PORT'));
+        $transport = new Swift_SmtpTransport(getenv('SMTP_SERVER'), (int)getenv('SMTP_PORT'));
     } else {
         return null;
     }
